Organizes all the test files in different packages
