+++ site/server/test/com/fortysevendeg/exercises/persistence/queries/UserQueriesSpec.scala
@@ -3,14 +3,16 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
+package com.fortysevendeg.exercises.persistence.queries
+
 import com.fortysevendeg.exercises.persistence.domain.{ UserQueries ⇒ Q }
 import com.fortysevendeg.exercises.persistence.repositories.UserRepository._
+import com.fortysevendeg.exercises.support.DatabaseInstance
 import doobie.contrib.specs2.analysisspec.AnalysisSpec
 import doobie.util.query.Query
 import doobie.util.update.Update
 import org.specs2.mutable.Specification
 import shared.User
-import test.database.DatabaseInstance
 
 class UserQueriesSpec
     extends Specification
